+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* [core] Add `#apply` for applying a list of operations (@janko-m)
+
 ## 1.0.0 (2018-04-04)
 
 * Depend on `mini_magick` and `ruby-vips` gems (@janko-m, @mokolabs)